--- test-toolkit/docker/1.15.2/Dockerfile.dlc.gpu
@@ -1,6 +1,0 @@
-ARG region
-FROM 763104351884.dkr.ecr.$region.amazonaws.com/tensorflow-training:1.15.2-gpu-py3
-
-COPY dist/sagemaker_tensorflow_training-*.tar.gz /sagemaker_tensorflow_training.tar.gz
-RUN pip install --upgrade --no-cache-dir /sagemaker_tensorflow_training.tar.gz && \
-    rm /sagemaker_tensorflow_training.tar.gz
