+++ test/integration/conftest.py
@@ -18,9 +18,10 @@ import os
 import boto3
 import pytest
 from sagemaker import LocalSession, Session
-from sagemaker.tensorflow import TensorFlow
 
-from test.integration import NO_P2_REGIONS, NO_P3_REGIONS
+from integration import image_utils
+from integration import NO_P2_REGIONS, NO_P3_REGIONS
+
 
 logger = logging.getLogger(__name__)
 logging.getLogger('boto').setLevel(logging.INFO)
@@ -29,16 +30,21 @@ logging.getLogger('factory.py').setLevel(logging.INFO)
 logging.getLogger('auth.py').setLevel(logging.INFO)
 logging.getLogger('connectionpool.py').setLevel(logging.INFO)
 
-SCRIPT_PATH = os.path.dirname(os.path.realpath(__file__))
+DIR_PATH = os.path.dirname(os.path.realpath(__file__))
 
 
 def pytest_addoption(parser):
-    parser.addoption('--docker-base-name', default='sagemaker-tensorflow-scriptmode')
+    parser.addoption('--build-image', '-B', action='store_true')
+    parser.addoption('--push-image', '-P', action='store_true')
+    parser.addoption('--dockerfile-type', '-T', choices=['dlc.cpu', 'dlc.gpu', 'tf'],
+                     default='tf')
+    parser.addoption('--dockerfile', '-D', default=None)
+    parser.addoption('--docker-base-name', default='sagemaker-tensorflow-training')
     parser.addoption('--tag', default=None)
     parser.addoption('--region', default='us-west-2')
-    parser.addoption('--framework-version', default=TensorFlow.LATEST_VERSION)
+    parser.addoption('--framework-version', default='1.15.2')
     parser.addoption('--processor', default='cpu', choices=['cpu', 'gpu', 'cpu,gpu'])
-    parser.addoption('--py-version', default='3', choices=['2', '3', '2,3', '37'])
+    parser.addoption('--py-version', default='3', choices=['2', '3', '2,3'])
     parser.addoption('--account-id', default='142577830533')
     parser.addoption('--instance-type', default=None)
 
@@ -46,6 +52,38 @@ def pytest_addoption(parser):
 def pytest_configure(config):
     os.environ['TEST_PY_VERSIONS'] = config.getoption('--py-version')
     os.environ['TEST_PROCESSORS'] = config.getoption('--processor')
+
+
+@pytest.fixture(scope='session', name='dockerfile_type')
+def fixture_dockerfile_type(request):
+    return request.config.getoption('--dockerfile-type')
+
+
+@pytest.fixture(scope='session', name='dockerfile')
+def fixture_dockerfile(request, dockerfile_type):
+    dockerfile = request.config.getoption('--dockerfile')
+    return dockerfile if dockerfile else 'Dockerfile.{}'.format(dockerfile_type)
+
+
+@pytest.fixture(scope='session', name='build_image', autouse=True)
+def fixture_build_image(request, framework_version, dockerfile, image_uri, region):
+    build_image = request.config.getoption('--build-image')
+    if build_image:
+        return image_utils.build_image(framework_version=framework_version,
+                                       dockerfile=dockerfile,
+                                       image_uri=image_uri,
+                                       region=region,
+                                       cwd=os.path.join(DIR_PATH, '..', '..'))
+
+    return image_uri
+
+
+@pytest.fixture(scope='session', name='push_image', autouse=True)
+def fixture_push_image(request, image_uri, region, account_id):
+    push_image = request.config.getoption('--push-image')
+    if push_image:
+        return image_utils.push_image(image_uri, region, account_id)
+    return None
 
 
 @pytest.fixture(scope='session')
@@ -63,7 +101,7 @@ def framework_version(request):
     return request.config.getoption('--framework-version')
 
 
-@pytest.fixture
+@pytest.fixture(scope='session')
 def tag(request, framework_version, processor, py_version):
     provided_tag = request.config.getoption('--tag')
     default_tag = '{}-{}-py{}'.format(framework_version, processor, py_version)
@@ -92,20 +130,6 @@ def instance_type(request, processor):
     return provided_instance_type if provided_instance_type is not None else default_instance_type
 
 
-@pytest.fixture()
-def py_version():
-    if 'TEST_PY_VERSIONS' in os.environ:
-        return os.environ['TEST_PY_VERSIONS'].split(',')
-    return None
-
-
-@pytest.fixture()
-def processor():
-    if 'TEST_PROCESSORS' in os.environ:
-        return os.environ['TEST_PROCESSORS'].split(',')
-    return None
-
-
 @pytest.fixture(autouse=True)
 def skip_by_device_type(request, processor):
     is_gpu = (processor == 'gpu')
@@ -121,19 +145,20 @@ def skip_gpu_instance_restricted_regions(region, instance_type):
         pytest.skip('Skipping GPU test in region {}'.format(region))
 
 
-@pytest.fixture
-def docker_image(docker_base_name, tag):
-    return '{}:{}'.format(docker_base_name, tag)
-
-
 @pytest.fixture(autouse=True)
-def skip_py2_containers(request, tag):
-    if request.node.get_closest_marker('skip_py2_containers'):
-        if 'py2' in tag:
-            pytest.skip('Skipping python2 container with tag {}'.format(tag))
+def skip_by_dockerfile_type(request, dockerfile_type):
+    is_generic = (dockerfile_type == 'tf')
+    if request.node.get_closest_marker('skip_generic') and is_generic:
+        pytest.skip('Skipping because running generic image without mpi and horovod')
 
 
-@pytest.fixture
-def ecr_image(account_id, docker_base_name, tag, region):
-    return '{}.dkr.ecr.{}.amazonaws.com/{}:{}'.format(
-        account_id, region, docker_base_name, tag)
+@pytest.fixture(name='docker_registry', scope='session')
+def fixture_docker_registry(account_id, region):
+    return '{}.dkr.ecr.{}.amazonaws.com'.format(account_id, region) if account_id else None
+
+
+@pytest.fixture(name='image_uri', scope='session')
+def fixture_image_uri(docker_registry, docker_base_name, tag):
+    if docker_registry:
+        return '{}/{}:{}'.format(docker_registry, docker_base_name, tag)
+    return '{}:{}'.format(docker_base_name, tag)
