+++ test/integration/local/test_training.py
@@ -18,7 +18,7 @@ import tarfile
 import pytest
 from sagemaker.tensorflow import TensorFlow
 
-from test.integration.utils import processor, py_version  # noqa: F401
+from integration.utils import processor, py_version  # noqa: F401
 
 RESOURCE_PATH = os.path.join(os.path.dirname(__file__), '..', '..', 'resources')
 TF_CHECKPOINT_FILES = ['graph.pbtxt', 'model.ckpt-0.index', 'model.ckpt-0.meta']
@@ -28,40 +28,18 @@ TF_CHECKPOINT_FILES = ['graph.pbtxt', 'model.ckpt-0.index', 'model.ckpt-0.meta']
 def py_full_version(py_version):  # noqa: F811
     if py_version == '2':
         return '2.7'
-    elif py_version == '37':
-        return '3.7'
     else:
         return '3.6'
 
 
 @pytest.mark.skip_gpu
-def test_py_versions(sagemaker_local_session, docker_image, py_full_version, framework_version, tmpdir):
-    output_path = 'file://{}'.format(tmpdir)
-    run_tf_training(script=os.path.join(RESOURCE_PATH, 'test_py_version', 'entry.py'),
-                    instance_type='local',
-                    instance_count=1,
-                    sagemaker_local_session=sagemaker_local_session,
-                    docker_image=docker_image,
-                    framework_version=framework_version,
-                    output_path=output_path,
-                    training_data_path=None)
-
-    with tarfile.open(os.path.join(str(tmpdir), 'output.tar.gz')) as tar:
-        output_file = tar.getmember('py_version')
-        tar.extractall(path=str(tmpdir), members=[output_file])
-
-    with open(os.path.join(str(tmpdir), 'py_version')) as f:
-        assert f.read().strip()[0] == py_full_version[0]
-
-
-@pytest.mark.skip_gpu
-def test_mnist_cpu(sagemaker_local_session, docker_image, tmpdir, framework_version):
+def test_mnist_cpu(sagemaker_local_session, image_uri, tmpdir, framework_version):
     output_path = 'file://{}'.format(tmpdir)
     run_tf_training(script=os.path.join(RESOURCE_PATH, 'mnist', 'mnist.py'),
                     instance_type='local',
                     instance_count=1,
                     sagemaker_local_session=sagemaker_local_session,
-                    docker_image=docker_image,
+                    image_uri=image_uri,
                     framework_version=framework_version,
                     output_path=output_path,
                     training_data_path='file://{}'.format(
@@ -69,21 +47,9 @@ def test_mnist_cpu(sagemaker_local_session, docker_image, tmpdir, framework_vers
     _assert_files_exist_in_tar(output_path, ['my_model.h5'])
 
 
-@pytest.mark.skip_cpu
-def test_gpu(sagemaker_local_session, docker_image, framework_version):
-    run_tf_training(script=os.path.join(RESOURCE_PATH, 'gpu_device_placement.py'),
-                    instance_type='local_gpu',
-                    instance_count=1,
-                    sagemaker_local_session=sagemaker_local_session,
-                    docker_image=docker_image,
-                    framework_version=framework_version,
-                    training_data_path='file://{}'.format(
-                        os.path.join(RESOURCE_PATH, 'mnist', 'data')))
-
-
 @pytest.mark.skip_gpu
 def test_distributed_training_cpu_no_ps(sagemaker_local_session,
-                                        docker_image,
+                                        image_uri,
                                         tmpdir,
                                         framework_version):
     output_path = 'file://{}'.format(tmpdir)
@@ -91,7 +57,7 @@ def test_distributed_training_cpu_no_ps(sagemaker_local_session,
                     instance_type='local',
                     instance_count=2,
                     sagemaker_local_session=sagemaker_local_session,
-                    docker_image=docker_image,
+                    image_uri=image_uri,
                     framework_version=framework_version,
                     output_path=output_path,
                     training_data_path='file://{}'.format(
@@ -101,7 +67,7 @@ def test_distributed_training_cpu_no_ps(sagemaker_local_session,
 
 @pytest.mark.skip_gpu
 def test_distributed_training_cpu_ps(sagemaker_local_session,
-                                     docker_image,
+                                     image_uri,
                                      tmpdir,
                                      framework_version):
     output_path = 'file://{}'.format(tmpdir)
@@ -109,7 +75,7 @@ def test_distributed_training_cpu_ps(sagemaker_local_session,
                     instance_type='local',
                     instance_count=2,
                     sagemaker_local_session=sagemaker_local_session,
-                    docker_image=docker_image,
+                    image_uri=image_uri,
                     framework_version=framework_version,
                     output_path=output_path,
                     hyperparameters={'sagemaker_parameter_server_enabled': True},
@@ -122,7 +88,7 @@ def run_tf_training(script,
                     instance_type,
                     instance_count,
                     sagemaker_local_session,
-                    docker_image,
+                    image_uri,
                     framework_version,
                     training_data_path,
                     output_path=None,
@@ -135,7 +101,7 @@ def run_tf_training(script,
                            train_instance_count=instance_count,
                            train_instance_type=instance_type,
                            sagemaker_session=sagemaker_local_session,
-                           image_name=docker_image,
+                           image_name=image_uri,
                            model_dir='/opt/ml/model',
                            output_path=output_path,
                            hyperparameters=hyperparameters,
