+++ test/resources/mnist/mnist_estimator.py
@@ -4,6 +4,8 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import logging
+
 import numpy as np
 import tensorflow as tf
 import os
@@ -22,7 +24,7 @@ def cnn_model_fn(features, labels, mode):
   # Padding is added to preserve width and height.
   # Input Tensor Shape: [batch_size, 28, 28, 1]
   # Output Tensor Shape: [batch_size, 28, 28, 32]
-  conv1 = tf.layers.conv2d(
+  conv1 = tf.compat.v1.layers.conv2d(
       inputs=input_layer,
       filters=32,
       kernel_size=[5, 5],
@@ -33,14 +35,14 @@ def cnn_model_fn(features, labels, mode):
   # First max pooling layer with a 2x2 filter and stride of 2
   # Input Tensor Shape: [batch_size, 28, 28, 32]
   # Output Tensor Shape: [batch_size, 14, 14, 32]
-  pool1 = tf.layers.max_pooling2d(inputs=conv1, pool_size=[2, 2], strides=2)
+  pool1 = tf.compat.v1.layers.max_pooling2d(inputs=conv1, pool_size=[2, 2], strides=2)
 
   # Convolutional Layer #2
   # Computes 64 features using a 5x5 filter.
   # Padding is added to preserve width and height.
   # Input Tensor Shape: [batch_size, 14, 14, 32]
   # Output Tensor Shape: [batch_size, 14, 14, 64]
-  conv2 = tf.layers.conv2d(
+  conv2 = tf.compat.v1.layers.conv2d(
       inputs=pool1,
       filters=64,
       kernel_size=[5, 5],
@@ -51,7 +53,7 @@ def cnn_model_fn(features, labels, mode):
   # Second max pooling layer with a 2x2 filter and stride of 2
   # Input Tensor Shape: [batch_size, 14, 14, 64]
   # Output Tensor Shape: [batch_size, 7, 7, 64]
-  pool2 = tf.layers.max_pooling2d(inputs=conv2, pool_size=[2, 2], strides=2)
+  pool2 = tf.compat.v1.layers.max_pooling2d(inputs=conv2, pool_size=[2, 2], strides=2)
 
   # Flatten tensor into a batch of vectors
   # Input Tensor Shape: [batch_size, 7, 7, 64]
@@ -62,16 +64,16 @@ def cnn_model_fn(features, labels, mode):
   # Densely connected layer with 1024 neurons
   # Input Tensor Shape: [batch_size, 7 * 7 * 64]
   # Output Tensor Shape: [batch_size, 1024]
-  dense = tf.layers.dense(inputs=pool2_flat, units=1024, activation=tf.nn.relu)
+  dense = tf.compat.v1.layers.dense(inputs=pool2_flat, units=1024, activation=tf.nn.relu)
 
   # Add dropout operation; 0.6 probability that element will be kept
-  dropout = tf.layers.dropout(
+  dropout = tf.compat.v1.layers.dropout(
       inputs=dense, rate=0.4, training=mode == tf.estimator.ModeKeys.TRAIN)
 
   # Logits layer
   # Input Tensor Shape: [batch_size, 1024]
   # Output Tensor Shape: [batch_size, 10]
-  logits = tf.layers.dense(inputs=dropout, units=10)
+  logits = tf.compat.v1.layers.dense(inputs=dropout, units=10)
 
   predictions = {
       # Generate predictions (for PREDICT and EVAL mode)
@@ -84,19 +86,19 @@ def cnn_model_fn(features, labels, mode):
     return tf.estimator.EstimatorSpec(mode=mode, predictions=predictions)
 
   # Calculate Loss (for both TRAIN and EVAL modes)
-  loss = tf.losses.sparse_softmax_cross_entropy(labels=labels, logits=logits)
+  loss = tf.compat.v1.losses.sparse_softmax_cross_entropy(labels=labels, logits=logits)
 
   # Configure the Training Op (for TRAIN mode)
   if mode == tf.estimator.ModeKeys.TRAIN:
-    optimizer = tf.train.GradientDescentOptimizer(learning_rate=0.001)
+    optimizer = tf.compat.v1.train.GradientDescentOptimizer(learning_rate=0.001)
     train_op = optimizer.minimize(
         loss=loss,
-        global_step=tf.train.get_global_step())
+        global_step=tf.compat.v1.train.get_global_step())
     return tf.estimator.EstimatorSpec(mode=mode, loss=loss, train_op=train_op)
 
   # Add evaluation metrics (for EVAL mode)
   eval_metric_ops = {
-      "accuracy": tf.metrics.accuracy(
+      "accuracy": tf.compat.v1.metrics.accuracy(
           labels=labels, predictions=predictions["classes"])}
   return tf.estimator.EstimatorSpec(
       mode=mode, loss=loss, eval_metric_ops=eval_metric_ops)
@@ -125,7 +127,7 @@ def _parse_args():
     return parser.parse_known_args()
 
 def serving_input_fn():
-    inputs = {'x': tf.placeholder(tf.float32, [None, 784])}
+    inputs = {'x': tf.compat.v1.placeholder(tf.float32, [None, 784])}
     return tf.estimator.export.ServingInputReceiver(inputs, inputs)
 
 if __name__ == "__main__":
@@ -133,7 +135,9 @@ if __name__ == "__main__":
     for arg in vars(args):
         print(arg, getattr(args, arg))
 
-    tf.logging.set_verbosity(tf.logging.DEBUG)
+    logger = tf.get_logger()
+    logger.setLevel(logging.DEBUG)
+    # tf.logging.set_verbosity(tf.logging.DEBUG)
     train_data, train_labels = _load_training_data(args.train)
     eval_data, eval_labels = _load_testing_data(args.train)
 
@@ -145,22 +149,22 @@ if __name__ == "__main__":
     # Set up logging for predictions
     # Log the values in the "Softmax" tensor with label "probabilities"
     tensors_to_log = {"probabilities": "softmax_tensor"}
-    logging_hook = tf.train.LoggingTensorHook(
+    logging_hook = tf.estimator.LoggingTensorHook(
         tensors=tensors_to_log, every_n_iter=50
     )
 
     # Train the model
-    train_input_fn = tf.estimator.inputs.numpy_input_fn(
+    train_input_fn = tf.compat.v1.estimator.inputs.numpy_input_fn(
         x={"x": train_data},
         y=train_labels,
         batch_size=args.batch_size,
         num_epochs=None,
         shuffle=True)
 
-    exporter = tf.estimator.LatestExporter('Servo', serving_input_receiver_fn=serving_input_fn) \
+    exporter = tf.compat.v1.estimator.LatestExporter('Servo', serving_input_receiver_fn=serving_input_fn) \
         if args.export_model_during_training else None
     # Evaluate the model and print results
-    eval_input_fn = tf.estimator.inputs.numpy_input_fn(
+    eval_input_fn = tf.compat.v1.estimator.inputs.numpy_input_fn(
         x={"x": eval_data},
         y=eval_labels,
         num_epochs=1,
@@ -171,4 +175,4 @@ if __name__ == "__main__":
     tf.estimator.train_and_evaluate(mnist_classifier, train_spec, eval_spec)
 
     if args.current_host == args.hosts[0]:
-        mnist_classifier.export_savedmodel('/opt/ml/model', serving_input_fn)
+        mnist_classifier.export_saved_model('/opt/ml/model', serving_input_fn)
